+++ packages/tailwindcss/src/intellisense.test.ts
@@ -21,6 +21,9 @@ function loadDesignSystem() {
   theme.add('--perspective-dramatic', '100px')
   theme.add('--perspective-normal', '500px')
   theme.add('--opacity-background', '0.3')
+  theme.add('--drop-shadow-sm', '0 1px 1px rgb(0 0 0 / 0.05)')
+  theme.add('--inset-shadow-sm', 'inset 0 1px 1px rgb(0 0 0 / 0.05)')
+  theme.add('--inset-ring-big', '100px')
   return buildDesignSystem(theme)
 }