+++ packages/tailwindcss/src/utilities.ts
@@ -23,6 +23,7 @@ type SuggestionDefinition =
       values?: string[]
       modifiers?: string[]
       valueThemeKeys?: ThemeKey[]
+      ignoredThemeKeys?: ThemeKey[]
       modifierThemeKeys?: ThemeKey[]
       hasDefaultValue?: boolean
     }
@@ -203,8 +204,8 @@ export function createUtilities(theme: Theme) {
    * Register list of suggestions for a class
    */
   function suggest(classRoot: string, defns: () => SuggestionDefinition[]) {
-    function* resolve(themeKeys: ThemeKey[]) {
-      for (let value of theme.keysInNamespaces(themeKeys)) {
+    function* resolve(themeKeys: ThemeKey[], ignoredThemeKeys: ThemeKey[] = []) {
+      for (let value of theme.keysInNamespaces(themeKeys, ignoredThemeKeys)) {
         yield value.replaceAll('_', '.')
       }
     }
@@ -220,7 +221,7 @@ export function createUtilities(theme: Theme) {
 
         let values: (string | null)[] = [
           ...(defn.values ?? []),
-          ...resolve(defn.valueThemeKeys ?? []),
+          ...resolve(defn.valueThemeKeys ?? [], defn.ignoredThemeKeys),
         ]
         let modifiers = [...(defn.modifiers ?? []), ...resolve(defn.modifierThemeKeys ?? [])]
 
@@ -252,6 +253,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative?: boolean
     supportsFractions?: boolean
     themeKeys?: ThemeKey[]
+    ignoredThemeKeys?: ThemeKey[]
     defaultValue?: string | null
     handleBareValue?: (value: NamedUtilityValue) => string | null
     handle: (value: string) => AstNode[] | undefined
@@ -277,7 +279,8 @@ export function createUtilities(theme: Theme) {
         // `defaultValue` (for candidates like `grow` that have no theme values)
         // or a bare theme value (like `--radius` for `rounded`). No utility
         // will ever support both of these.
-        value = desc.defaultValue ?? theme.resolve(null, desc.themeKeys ?? [])
+        value =
+          desc.defaultValue ?? theme.resolve(null, desc.themeKeys ?? [], desc.ignoredThemeKeys)
       } else if (candidate.value.kind === 'arbitrary') {
         if (candidate.modifier) return
         value = candidate.value.value
@@ -285,6 +288,7 @@ export function createUtilities(theme: Theme) {
         value = theme.resolve(
           candidate.value.fraction ?? candidate.value.value,
           desc.themeKeys ?? [],
+          desc.ignoredThemeKeys,
         )
 
         // Automatically handle things like `w-1/2` without requiring `1/2` to
@@ -314,6 +318,7 @@ export function createUtilities(theme: Theme) {
       {
         supportsNegative: desc.supportsNegative,
         valueThemeKeys: desc.themeKeys ?? [],
+        ignoredThemeKeys: desc.ignoredThemeKeys ?? [],
         hasDefaultValue: desc.defaultValue !== undefined && desc.defaultValue !== null,
       },
     ])
@@ -424,49 +429,13 @@ export function createUtilities(theme: Theme) {
     let value = candidate.negative ? '-100%' : '100%'
     return [decl('inset', value)]
   })
-  utilities.functional('inset', (candidate) => {
-    if (!candidate.value) return
-
-    let value
-    if (candidate.value.kind === 'arbitrary') {
-      if (candidate.modifier) return
-      value = candidate.value.value
-    } else {
-      // We need to make sure variables like `--inset-shadow-sm` and
-      // `--inset-ring-thick` don't mistakenly generate utilities for the
-      // `inset` property.
-      if (
-        candidate.value.value === 'ring' ||
-        candidate.value.value === 'shadow' ||
-        candidate.value.value.startsWith('ring-') ||
-        candidate.value.value.startsWith('shadow-')
-      ) {
-        value = theme.resolve(candidate.value.fraction ?? candidate.value.value, ['--spacing'])
-      } else {
-        value = theme.resolve(candidate.value.fraction ?? candidate.value.value, [
-          '--inset',
-          '--spacing',
-        ])
-      }
-
-      if (!value && candidate.value.fraction) {
-        let [lhs, rhs] = segment(candidate.value.fraction, '/')
-        if (!isPositiveInteger(lhs) || !isPositiveInteger(rhs)) return
-        value = `calc(${candidate.value.fraction} * 100%)`
-      }
-
-      if (!value) return
-    }
-    value = withNegative(value, candidate)
-    return [decl('inset', value)]
+  functionalUtility('inset', {
+    supportsNegative: true,
+    supportsFractions: true,
+    themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
+    handle: (value) => [decl('inset', value)],
   })
-  suggest('inset', () => [
-    {
-      supportsNegative: true,
-      valueThemeKeys: ['--inset', '--spacing'],
-      hasDefaultValue: false,
-    },
-  ])
 
   staticUtility('inset-x-auto', [['inset-inline', 'auto']])
 
@@ -478,6 +447,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('inset-inline', value)],
   })
 
@@ -490,6 +460,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('inset-block', value)],
   })
 
@@ -502,6 +473,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('inset-inline-start', value)],
   })
 
@@ -514,6 +486,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('inset-inline-end', value)],
   })
 
@@ -526,6 +499,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('top', value)],
   })
 
@@ -538,6 +512,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('right', value)],
   })
 
@@ -550,6 +525,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('bottom', value)],
   })
 
@@ -562,6 +538,7 @@ export function createUtilities(theme: Theme) {
     supportsNegative: true,
     supportsFractions: true,
     themeKeys: ['--inset', '--spacing'],
+    ignoredThemeKeys: ['--inset-ring', '--inset-shadow'],
     handle: (value) => [decl('left', value)],
   })
 
@@ -3655,6 +3632,11 @@ export function createUtilities(theme: Theme) {
       },
     ])
 
+    staticUtility('drop-shadow-none', [
+      filterProperties,
+      ['--tw-drop-shadow', ' '],
+      ['filter', cssFilterValue],
+    ])
     functionalUtility('drop-shadow', {
       themeKeys: ['--drop-shadow'],
       handle: (value) => [